wip: refactor(GC3): compare container sizes without int32_t casts
Both sides are size_t, so the paired static_casts added by the automated rewrite were noise.
Sites that subtract from a size (size() - 1, size() - 2) keep the signed cast: unsigned
arithmetic would wrap on an empty container.

+++ Desert/Desert/Source/Engine/Geometry/UECore/DynamicMesh/DynamicMesh3_Queries.cpp
@@ -302,9 +302,8 @@ EMeshResult FDynamicMesh3::GetVtxContiguousTriangles( int VertexID, IntArray& Tr
         SpanLengths.push_back( static_cast<int32_t>( TrianglesOut.size() ) - SpanStart );
     }
 
-    return UE_ENSURE( static_cast<int32_t>( SpanLengths.size() ) == static_cast<int32_t>( IsLoop.size() ) )
-                ? EMeshResult::Ok
-                : EMeshResult::Failed_InvalidNeighbourhood;
+    return UE_ENSURE( SpanLengths.size() == IsLoop.size() ) ? EMeshResult::Ok
+                                                            : EMeshResult::Failed_InvalidNeighbourhood;
 }
 
 template EMeshResult FDynamicMesh3::GetVtxContiguousTriangles<std::vector<int32_t>, std::vector<bool>>(
